Refactoring
In this commit I attempted to refactor the code in directions.jsx
I was able to extract two main components, the 1st one being the LoactionSelector and the 2nd  was the ModalSearchBars.
